+++ src/components/App.js
@@ -5,7 +5,6 @@ import { Helmet } from 'react-helmet';
 import { useSession } from 'hooks';
 import RouteFromPath from 'components/routes/RouteFromPath';
 import routes from '../routes';
-import Header from './common/Header';
 
 const App = () => {
   const { authenticated } = useSession();
@@ -15,7 +14,6 @@ const App = () => {
       <Helmet>
         <title>RS React Redux Base</title>
       </Helmet>
-      <Header />
       <BrowserRouter>
         <Switch>
           {routes.map((route, index) => (
