+++ src/components/common/Input.js
@@ -13,11 +13,15 @@ const Input = ({ label, name, value, onChange, errors, active, touched, ...props
 
   return (
     <div>
-      {label && <label htmlFor={name}>{label}</label>}
-      <div>
-        <input name={name} value={value} onChange={onChange} {...props} />
+      {label && (
+        <label htmlFor={name} className="input-label">
+          {label}
+        </label>
+      )}
+      <div className="input-container">
+        <input name={name} value={value} onChange={onChange} {...props} className="input" />
         {touched && errors && (
-          <span>
+          <span className="span-error">
             <FormattedMessage
               id={parseInputErrors(errors)}
               defaultMessage={parseInputErrors(errors)}
