+++ src/pages/loginPage/loginPage.scss
@@ -1,0 +1,20 @@
+.signup-button {
+  color: #000;
+  line-height: 15px;
+  font-size: 11px;
+  text-transform: uppercase;
+}
+
+.selected {
+  color: #000;
+}
+
+.social-buttons-container {
+  margin: 20px 0;
+  width: 100px;
+  justify-content: space-around;
+}
+
+.social-button-flex {
+  justify-content: space-between;
+}
