+++ src/routes.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import routesPaths from 'constants/routesPaths';
 import HomePage from 'pages/HomePage';
-import LoginPage from 'pages/LoginPage';
+import LoginPage from 'pages/loginPage/LoginPage';
 import SignUpPage from 'pages/SignUpPage';
 import NotFoundPage from 'pages/NotFoundPage';
