+++ src/styles/_variables.scss
@@ -1,2 +1,4 @@
+@import url('https://fonts.googleapis.com/css?family=Open+Sans');
+
 // Typography
-$font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
+$font-family: 'Open Sans', sans-serif;
